+++ src/components/agent/AgentWorkspace.tsx
@@ -19,9 +19,7 @@ import {
   hermesBrowserAccess,
   primeGeneratedVideoDir,
   importHermesBridgeFileBytes,
-  listVeniceModels,
   downloadHermesBridgeFile,
-  providerModelSettings,
   revealPath,
   type AgentTaskStatus,
   type HermesSessionMessage,
@@ -56,10 +54,6 @@ import {
   // `AgentArtifact` (the file-viewer card), so alias it.
   type AgentArtifact as TimelineArtifact,
 } from "../../lib/hermes-artifact-store";
-import {
-  PROVIDER_MODEL_SETTINGS_CHANGED_EVENT,
-  type ProviderModelSettingsChangedDetail,
-} from "../../lib/model-privacy";
 import { localGenerationOptionId } from "../../lib/local-generation";
 import {
   rememberSessionThinkingLevel,
@@ -98,6 +92,7 @@ import {
 } from "../../lib/agent-chat-gallery";
 import { attachScrollThumbFade } from "../../lib/scroll-thumb-fade";
 import type { AgentWorkspaceProps } from "./agent-workspace-types";
+import { useAgentModelSelection } from "./use-agent-model-selection";
 import { createComposerDispatchActions } from "./composer-dispatch-actions";
 import { createIssueReportStateActions } from "./issue-report-state-actions";
 import { createAttachmentImportActions } from "./attachment-import-actions";
@@ -106,7 +101,6 @@ import { useAgentSessionLoading } from "./use-agent-session-loading";
 import { useAgentSelection } from "./use-agent-selection";
 import { useAgentRuntimeState } from "./use-agent-runtime-state";
 import { useAgentCoreState } from "./use-agent-core-state";
-import { createModelSelectionActions } from "./model-selection-actions";
 import { createPendingImageActions } from "./pending-image-actions";
 import { createIssueReportActions } from "./issue-report-actions";
 import { createComposerFileEvents } from "./composer/composer-file-events";
@@ -1419,128 +1413,29 @@ export function AgentWorkspace({
   // active provider, the saved local endpoint, and the pill selection (the
   // synthetic local option when local is active). Shared by the mount fetch
   // and the model-switch handler so both stay in lockstep with the backend.
-  const commitGenerationSettings = useCallback(
-    (settings: ProviderModelSettingsDto, fallbackModelId = "") => {
-      const local = settings.localGeneration ?? {
-        baseUrl: "",
-        modelId: "",
-        apiKey: "",
-      };
-      const selectedModelId = generationSelectionId(settings, fallbackModelId);
-      localGenerationRef.current = local;
-      setLocalGeneration(local);
-      defaultGenerationModelIdRef.current = selectedModelId;
-      setDefaultGenerationModelId(selectedModelId);
-      confirmedCostQualityRef.current = settings.costQuality;
-      generationCostQualityRef.current = settings.costQuality;
-      setGenerationCostQuality(settings.costQuality);
-      veniceApiKeyConfiguredRef.current = settings.veniceApiKeyConfigured;
-      setVeniceApiKeyConfigured(settings.veniceApiKeyConfigured);
-      return selectedModelId;
-    },
-    [],
-  );
-
-  // Out-of-order responses (a slow mount fetch landing after a settings
-  // change refresh) must not clobber the newer result.
-  const generationModelRequestSequence = useRef(0);
-  const loadGenerationModel = useCallback(async () => {
-    const requestId = ++generationModelRequestSequence.current;
-    try {
-      const settingsPromise = providerModelSettings();
-      const modelsPromise = listVeniceModels("generation");
-      // Surfaced before the catalog await: the settings read is local IPC, so
-      // key-presence state (the Auto billing note) refreshes even when the
-      // remote catalog fetch fails.
-      modelsPromise.catch(() => {});
-      const settingsResponse = await settingsPromise;
-      if (requestId === generationModelRequestSequence.current) {
-        veniceApiKeyConfiguredRef.current = settingsResponse.settings.veniceApiKeyConfigured;
-        setVeniceApiKeyConfigured(settingsResponse.settings.veniceApiKeyConfigured);
-      }
-      const modelsResponse = await modelsPromise;
-      const selectedModelId = generationSelectionId(
-        settingsResponse.settings,
-        modelsResponse.selectedModel,
-      );
-      if (requestId === generationModelRequestSequence.current) {
-        generationModelsRef.current = modelsResponse.models;
-        setGenerationModels(modelsResponse.models);
-        commitGenerationSettings(settingsResponse.settings, modelsResponse.selectedModel);
-      }
-      return { models: modelsResponse.models, selectedModelId };
-    } catch {
-      if (requestId === generationModelRequestSequence.current) {
-        defaultGenerationModelIdRef.current = "";
-        generationModelsRef.current = [];
-        setDefaultGenerationModelId("");
-      }
-      return null;
-    }
-  }, [commitGenerationSettings]);
-
-  useEffect(() => {
-    defaultGenerationModelIdRef.current = defaultGenerationModelId;
-    const defaultModelId = defaultGenerationModelId.trim();
-    if (!defaultModelId) return;
-    setHermesSessionItems((current) => {
-      let changed = false;
-      const next = current.map((session) => {
-        if (session.model?.trim()) return session;
-        changed = true;
-        return { ...session, model: defaultModelId };
-      });
-      return changed ? next : current;
-    });
-  }, [defaultGenerationModelId]);
-
-  useEffect(() => {
-    function handleProviderModelSettingsChanged(event: Event) {
-      const { mode } = (event as CustomEvent<ProviderModelSettingsChangedDetail>).detail;
-      if (mode === "generation") {
-        void loadGenerationModel();
-      }
-    }
-
-    void loadGenerationModel();
-    window.addEventListener(
-      PROVIDER_MODEL_SETTINGS_CHANGED_EVENT,
-      handleProviderModelSettingsChanged,
-    );
-    return () => {
-      window.removeEventListener(
-        PROVIDER_MODEL_SETTINGS_CHANGED_EVENT,
-        handleProviderModelSettingsChanged,
-      );
-    };
-  }, [loadGenerationModel]);
-
   const {
+    loadGenerationModel,
     commitSessionModelSelections,
-    storedSessionIdForComposerModelSelection,
-    queueComposerSessionModelSelection,
     captureSessionModelTarget,
     openComposerModelPicker,
-    markRemoteGenerationSelected,
     saveGenerationSelection,
-    selectLocalGeneration,
     handleCostQualityChange,
     handleSelectGenerationModel,
-  } = createModelSelectionActions({
+  } = useAgentModelSelection({
     MODEL_SWITCH_TOAST_ID,
     activeGenerationCostQuality,
     confirmedCostQualityRef,
     costQualitySaveChainRef,
+    defaultGenerationModelId,
     defaultGenerationModelIdRef,
-    generationCostQuality,
     generationCostQualityRef,
-    generationModelRequestSequence,
+    generationCostQuality,
     generationModelsRef,
+    generationSelectionId,
     generationSelectionIntentRevisionRef,
     generationSelectionSaveChainRef,
     hermesSessionItemsRef,
     latestCostQualitySaveRef,
-    loadGenerationModel,
     localEnableConfirmArmedForRef,
     localGenerationRef,
     newSessionModeRef,
@@ -1551,12 +1446,17 @@ export function AgentWorkspace({
     setComposerModelFromSlash,
     setComposerModelOpen,
     setDefaultGenerationModelId,
-    setError,
     setGenerationCostQuality,
+    setGenerationModels,
+    setHermesSessionItems,
+    setLocalGeneration,
+    setError,
     setModelRootSearch,
     setModelSearch,
     setSandboxMenuOpen,
     setSessionModelSelections,
+    setVeniceApiKeyConfigured,
+    veniceApiKeyConfiguredRef,
   });
 
   useEffect(() => {
